+++ controllers/operator/appdbreplicaset_controller.go
@@ -1046,12 +1046,22 @@ func (r *ReconcileAppDbReplicaSet) buildAppDbAutomationConfig(ctx context.Contex
 				Destination: automationconfig.File,
 				Path:        path.Join(util.PvcMountPathLogs, "mongodb.log"),
 			}
+
 			if opsManager.Spec.AppDB.AutomationAgent.SystemLog != nil {
 				systemLog = opsManager.Spec.AppDB.AutomationAgent.SystemLog
 			}
 
+			// This setting takes precedence, above has been deprecated, and we should favor the one after mongod
+			if opsManager.Spec.AppDB.AutomationAgent.Mongod.SystemLog != nil {
+				systemLog = opsManager.Spec.AppDB.AutomationAgent.Mongod.SystemLog
+			}
+
 			if acType == automation {
-				automationconfig.ConfigureAgentConfiguration(systemLog, opsManager.Spec.AppDB.AutomationAgent.LogRotate, p)
+				if opsManager.Spec.AppDB.AutomationAgent.Mongod.HasLoggingConfigured() {
+					automationconfig.ConfigureAgentConfiguration(systemLog, opsManager.Spec.AppDB.AutomationAgent.Mongod.LogRotate, opsManager.Spec.AppDB.AutomationAgent.Mongod.AuditLogRotate, p)
+				} else {
+					automationconfig.ConfigureAgentConfiguration(systemLog, opsManager.Spec.AppDB.AutomationAgent.LogRotate, opsManager.Spec.AppDB.AutomationAgent.Mongod.AuditLogRotate, p)
+				}
 			}
 		}).
 		AddModifications(func(automationConfig *automationconfig.AutomationConfig) {